:truck: Split IR models into multiple directories
